fastapi/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
authlib==1.6.5
httpx==0.28.1
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
neo4j==5.25.0

